requirements.txt
pyyaml
robotframework
robotframework-seleniumlibrary